+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+  * Replace rustls-pemfile usage with rustls-pki-types #1122
   * Support for env NO_PROXY and proxy config #1118
 
 # 3.1.2